Add python 3 compatibility
