+++ pydecorate/__init__.py
@@ -16,7 +16,7 @@
 #You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from DecoratorAGG import DecoratorAGG
+from pydecorate.decorator_agg import DecoratorAGG
