+++ pydecorate/decorator_agg.py
@@ -15,12 +15,12 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from pydecorate.DecoratorBase import DecoratorBase
+from pydecorate.decorator_base import DecoratorBase
 
 try:
     from PIL import ImageDraw
 except ImportError:
-    print "ImportError: Missing module: ImageDraw"
+    print("ImportError: Missing module: ImageDraw")
 
 
 class DecoratorAGG(DecoratorBase):
